examples/python-usage/pyproject.toml
# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".

[build-system]
requires = [ "poetry_core>=1.0.0" ]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "python-usage"
version = "0.1.0"
description = ""
authors = [ "Galen Dunkleberger <awsgalen@amazon.com>" ]
readme = "README.md"

  [tool.poetry.dependencies]
  aws-cdk-lib = "^2.65.0"
  boto3 = "*"
  constructs = "^10.0.5"
  python = "^3.7"

  [tool.poetry.dev-dependencies]
  projen = "0.67.60"
